feat(tier-lists): wrap items, responsive sizing, shared color picker
- Tier rows now wrap items instead of horizontal-scrolling — no more off-screen hidden cards on narrow screens; live view now matches the PNG export layout
- Responsive tier sizing (TierRowMetrics): on <500px screens cards shrink to 64×86 and the tier label column to 48px; unranked pool uses the same metrics; row minHeight stays ~12px taller than the card
- Extract ColorPickerDialog into a shared widget (palette + HSL sliders + hex preview); use it in tags, tier lists, profile create/edit, and canvas connection edit
- Add Profile.colorToHex helper; drop 11 orphan/duplicate l10n keys (9 connection color names + redundant profileColor/connectionColorLabel in favor of colorPickerTitle)
