+++ lib/l10n/app_en.arb
@@ -682,18 +682,8 @@
   "linkLabelOptional": "Label (optional)",
   "linkLabelHint": "My Link",
 
-  "connectionColorGray": "Gray",
-  "connectionColorRed": "Red",
-  "connectionColorOrange": "Orange",
-  "connectionColorYellow": "Yellow",
-  "connectionColorGreen": "Green",
-  "connectionColorBlue": "Blue",
-  "connectionColorPurple": "Purple",
-  "connectionColorBlack": "Black",
-  "connectionColorWhite": "White",
   "editConnectionTitle": "Edit Connection",
   "connectionLabelHint": "e.g. depends on, related to...",
-  "connectionColorLabel": "Color",
   "connectionStyleLabel": "Style",
   "connectionStyleSolid": "Solid",
   "connectionStyleDashed": "Dashed",
@@ -1537,7 +1527,6 @@
   },
   "cannotDeleteLastProfile": "Cannot delete the last profile",
   "profileName": "Name",
-  "profileColor": "Color",
   "whoIsPlayingToday": "Who's playing today?",
   "dontAskAgain": "Don't ask again",
   "profileStats": "{collections} collections, {items} items",
