+++ lib/l10n/app_localizations_ru.dart
@@ -1510,40 +1510,10 @@ class SRu extends S {
   String get linkLabelHint => 'Моя ссылка';
 
   @override
-  String get connectionColorGray => 'Серый';
-
-  @override
-  String get connectionColorRed => 'Красный';
-
-  @override
-  String get connectionColorOrange => 'Оранжевый';
-
-  @override
-  String get connectionColorYellow => 'Жёлтый';
-
-  @override
-  String get connectionColorGreen => 'Зелёный';
-
-  @override
-  String get connectionColorBlue => 'Синий';
-
-  @override
-  String get connectionColorPurple => 'Фиолетовый';
-
-  @override
-  String get connectionColorBlack => 'Чёрный';
-
-  @override
-  String get connectionColorWhite => 'Белый';
-
-  @override
   String get editConnectionTitle => 'Редактировать соединение';
 
   @override
   String get connectionLabelHint => 'напр. зависит от, связано с...';
-
-  @override
-  String get connectionColorLabel => 'Цвет';
 
   @override
   String get connectionStyleLabel => 'Стиль';
@@ -3333,9 +3303,6 @@ class SRu extends S {
 
   @override
   String get profileName => 'Имя';
-
-  @override
-  String get profileColor => 'Цвет';
 
   @override
   String get whoIsPlayingToday => 'Кто сегодня играет?';
