+++ lib/l10n/app_ru.arb
@@ -545,18 +545,8 @@
   "linkLabelOptional": "Подпись (необязательно)",
   "linkLabelHint": "Моя ссылка",
 
-  "connectionColorGray": "Серый",
-  "connectionColorRed": "Красный",
-  "connectionColorOrange": "Оранжевый",
-  "connectionColorYellow": "Жёлтый",
-  "connectionColorGreen": "Зелёный",
-  "connectionColorBlue": "Синий",
-  "connectionColorPurple": "Фиолетовый",
-  "connectionColorBlack": "Чёрный",
-  "connectionColorWhite": "Белый",
   "editConnectionTitle": "Редактировать соединение",
   "connectionLabelHint": "напр. зависит от, связано с...",
-  "connectionColorLabel": "Цвет",
   "connectionStyleLabel": "Стиль",
   "connectionStyleSolid": "Сплошная",
   "connectionStyleDashed": "Пунктирная",
@@ -1262,7 +1252,6 @@
   "deleteProfileConfirm": "Удалить профиль {name}? Все коллекции, вишлист и настройки будут удалены. Это действие нельзя отменить.",
   "cannotDeleteLastProfile": "Нельзя удалить последний профиль",
   "profileName": "Имя",
-  "profileColor": "Цвет",
   "whoIsPlayingToday": "Кто сегодня играет?",
   "dontAskAgain": "Не спрашивать снова",
   "profileStats": "{collections,plural, =1{1 коллекция} few{{collections} коллекции} other{{collections} коллекций}}, {items,plural, =1{1 тайтл} few{{items} тайтла} other{{items} тайтлов}}",
